Un-nest musicxml.write module
